+++ src/main/java/net/geforcemods/securitycraft/ConfigHandler.java
@@ -7,6 +7,7 @@ import java.util.function.Supplier;
 
 import org.apache.commons.lang3.tuple.Pair;
 
+import net.geforcemods.securitycraft.misc.TintMode;
 import net.geforcemods.securitycraft.util.TeamUtils;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.effect.MobEffectInstance;
@@ -14,6 +15,7 @@ import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.common.ForgeConfigSpec.BooleanValue;
 import net.minecraftforge.common.ForgeConfigSpec.ConfigValue;
 import net.minecraftforge.common.ForgeConfigSpec.DoubleValue;
+import net.minecraftforge.common.ForgeConfigSpec.EnumValue;
 import net.minecraftforge.common.ForgeConfigSpec.IntValue;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
@@ -43,9 +45,9 @@ public class ConfigHandler {
 
 	public static class Client {
 		public final BooleanValue sayThanksMessage;
-		public final BooleanValue reinforcedBlockTint;
 		public final BooleanValue debugCameraResetTracing;
 		public final BooleanValue morePerformantSRIRendering;
+		public final EnumValue<TintMode> reinforcedBlockTintMode;
 		public final IntValue reinforcedBlockTintColor;
 		public final IntValue frameFeedRenderDistance;
 		public final IntValue frameFeedResolution;
@@ -57,10 +59,6 @@ public class ConfigHandler {
 					.comment("Send a welcome message containing tips when joining the world")
 					.define("sayThanksMessage", true);
 
-			reinforcedBlockTint = builder
-					.comment("Should reinforced blocks' textures be slightly darker than their vanilla counterparts? Servers can force this setting on clients if the server config setting \"force_reinforced_block_tint\" is set to true.")
-					.define("reinforced_block_tint", true);
-
 			debugCameraResetTracing = builder
 					.comment("If this debug feature is enabled, SecurityCraft will attempt to find and report mods that prevent the feature of viewing security cameras from working when they immediately reset the player's camera entity.")
 					.define("debug_camera_reset_tracing", false);
@@ -69,8 +67,12 @@ public class ConfigHandler {
 					.comment("Should Secure Redstone Interfaces in \"Receiver\" mode be rendered without the spinning disk? This may lead to slight clientside performance gains.")
 					.define("more_performant_sri_rendering", false);
 
+			reinforcedBlockTintMode = builder
+					.comment("Controls what condition placed reinforced blocks must satisfy in order to be colored compared to their vanilla counterparts. The color of this tint can be defined through the \"reinforced_block_tint_color\" configuration option. The tint may apply to all reinforced blocks, no reinforced blocks, or only the reinforced blocks that belong or don't belong to the player.")
+					.defineEnum("reinforced_block_tint_mode", TintMode.ALL);
+
 			reinforcedBlockTintColor = builder
-					.comment("Set the color that reinforced blocks' textures have when reinforced_block_tint is enabled. This cannot be overridden by servers, and will be applied the same to all blocks. Grayscale values look best.",
+					.comment("Set the color that reinforced blocks' textures are tinted with. This will be applied the same to all blocks that satisfy the condition set by \"reinforced_block_tint_mode\". Grayscale values look best.",
 							"Format: 0xRRGGBB")
 					.defineInRange("reinforced_block_tint_color", 0x999999, 0x000000, 0xFFFFFF);
 
@@ -100,8 +102,6 @@ public class ConfigHandler {
 		public final IntValue inventoryScannerRange;
 		public final IntValue maxAlarmRange;
 		public final BooleanValue allowBlockClaim;
-		public final BooleanValue reinforcedBlockTint;
-		public final BooleanValue forceReinforcedBlockTint;
 		public final BooleanValue retinalScannerFace;
 		public final BooleanValue enableTeamOwnership;
 		public final ConfigValue<List<? extends String>> teamOwnershipPrecedence;
@@ -176,14 +176,6 @@ public class ConfigHandler {
 			allowBlockClaim = builder
 					.comment("Allows to claim blocks that do not have an owner by rightclicking them with the Universal Owner Changer.")
 					.define("allowBlockClaim", false);
-
-			reinforcedBlockTint = builder
-					.comment("Should reinforced blocks' textures be slightly darker than their vanilla counterparts? Servers can force this setting on clients if the server config \"force_reinforced_block_tint\" is set to true.")
-					.define("reinforced_block_tint", true);
-
-			forceReinforcedBlockTint = builder
-					.comment("Set this to true if you want to force the setting of reinforced_block_tint for players.")
-					.define("force_reinforced_block_tint", false);
 
 			retinalScannerFace = builder
 					.comment("Display owner face on retinal scanner?")
@@ -326,6 +318,8 @@ public class ConfigHandler {
 			loadEffects(SERVER.poweredTaserEffectsValue, SERVER.poweredTaserEffects);
 			updateTeamPrecedence(event);
 		}
+		else if (event.getConfig().getSpec() == CLIENT_SPEC && CLIENT_SPEC.isLoaded())
+			loadTintSettingsFromConfig();
 	}
 
 	private static void loadEffects(ConfigValue<List<? extends String>> effectsValue, List<Supplier<MobEffectInstance>> effects) {
@@ -362,6 +356,11 @@ public class ConfigHandler {
 		}
 
 		return true;
+	}
+
+	public static void loadTintSettingsFromConfig() {
+		TintMode.setColor(ConfigHandler.CLIENT.reinforcedBlockTintColor.get());
+		TintMode.setMode(ConfigHandler.CLIENT.reinforcedBlockTintMode.get());
 	}
 
 	private static void updateTeamPrecedence(ModConfigEvent event) {