+++ src/main/java/net/geforcemods/securitycraft/SCEventHandler.java
@@ -53,6 +53,7 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
@@ -80,6 +81,7 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.SkullBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.NoteBlockInstrument;
+import net.minecraft.world.level.chunk.ChunkAccess;
 import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
@@ -111,6 +113,7 @@ import net.minecraftforge.network.PacketDistributor;
 public class SCEventHandler {
 	private static final Integer NOTE_DELAY = 9;
 	public static final Map<Player, MutablePair<Integer, Deque<NoteWrapper>>> PLAYING_TUNES = new HashMap<>();
+	public static final Map<ResourceKey<Level>, List<ChunkAccess>> TINT_UPDATE_QUEUE = new HashMap<>();
 
 	private SCEventHandler() {}
 
@@ -154,6 +157,15 @@ public class SCEventHandler {
 				while (entries.hasNext()) {
 					if (entries.next().getValue().left == -1)
 						entries.remove();
+				}
+			}
+
+			for (ResourceKey<Level> levelResourceKey : TINT_UPDATE_QUEUE.keySet()) {
+				List<ChunkAccess> chunksToRecompile = TINT_UPDATE_QUEUE.get(levelResourceKey);
+
+				if (!chunksToRecompile.isEmpty()) {
+					event.getServer().getLevel(levelResourceKey).getChunkSource().chunkMap.resendBiomesForChunks(chunksToRecompile); //Tells the client to mark all modified sections as dirty, to properly update block tints. /fillbiome uses this too
+					chunksToRecompile.clear();
 				}
 			}
 		}