+++ src/main/java/net/geforcemods/securitycraft/items/UniversalOwnerChangerItem.java
@@ -8,6 +8,7 @@ import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.api.Owner;
 import net.geforcemods.securitycraft.blockentities.DisplayCaseBlockEntity;
 import net.geforcemods.securitycraft.misc.ModuleType;
+import net.geforcemods.securitycraft.util.ClientUtils;
 import net.geforcemods.securitycraft.util.IBlockMine;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.geforcemods.securitycraft.util.Utils;
@@ -89,6 +90,8 @@ public class UniversalOwnerChangerItem extends Item {
 
 		if (!level.isClientSide)
 			level.sendBlockUpdated(pos, state, state, 3);
+		else
+			ClientUtils.recompileChunk(pos);
 
 		//disable this in a development environment
 		if (FMLEnvironment.production && be instanceof IModuleInventory inv) {